refactoring connection stuff
